+++ R/zzz.R
@@ -1,4 +1,4 @@
-##‘ @importFrom memoise memoise
+##' @importFrom memoise memoise
 .onLoad <- function(libname, pkgname) { 
     yread <<- memoise::memoise(yread)
     yread_tsv <<- memoise::memoise(yread_tsv)